Moduleify parseQueryString
