+++ app/assets/javascripts/student_profile/parse_query_string.js
@@ -1,0 +1,37 @@
+// Adapted from https://github.com/sindresorhus/query-string/blob/master/index.js#L8
+
+export default function parseQueryString (str) {
+  if (typeof str !== 'string') {
+    return {};
+  }
+
+  str = str.trim().replace(/^(\?|#|&)/, '');
+
+  if (!str) {
+    return {};
+  }
+
+  return str.split('&').reduce(function (ret, param) {
+    const parts = param.replace(/\+/g, ' ').split('=');
+    // Firefox (pre 40) decodes `%3D` to `=`
+    // https://github.com/sindresorhus/query-string/pull/37
+    let key = parts.shift();
+    let val = parts.length > 0 ? parts.join('=') : undefined;
+
+    key = decodeURIComponent(key);
+
+    // missing `=` should be `null`:
+    // http://w3.org/TR/2012/WD-url-20120524/#collect-url-parameters
+    val = val === undefined ? null : decodeURIComponent(val);
+
+    if (!ret.hasOwnProperty(key)) {
+      ret[key] = val;
+    } else if (Array.isArray(ret[key])) {
+      ret[key].push(val);
+    } else {
+      ret[key] = [ret[key], val];
+    }
+
+    return ret;
+  }, {});
+}
